bugfix: @(next) in function called with match-fun.
The match-fun function must augment the input list
with the current input source, because @(next)
pops the first item from the files list and tries
to open the second.

We want it so that if we invoke

  (match-fun 'f arglist input '("abc"))

then if the pattern function f invokes @(next),
it will open "abc".

* match.c (match_fun): Calculate a value for the curfile
property of the match context and pass it to mf_all.  If the
input is a stream, we get its name.  We also push this curfile
onto the files list, satisfying the expectation that curfile
and the first element of files refer to the same thing.