bugfix: neglected unbound warnings in DWIM syntax.
DWIM expressions like [a b c] are not raising expansion-time
warnings about a, b, c being unbound.

* eval.c (expand_lisp1): The problem is that here we just
return in the case that the symbol is bindable and has no
macro expansion! Before returning, we must check whether
the symbol has a binding in the variable or function space.
If not, raise a warning.