Windows: shortcuts should start in user profile dir.
Noticed that the shortcuts created by NSIS are configured
to start in the TXR standard library installation
directory. Why is that? Because the locatio specified
by the most recent SetOutPath is used for this!

* inst.nsi: Before creating shortcuts, switch
SetOutPath to $PROFILE, which is NSIS's variable denoting
the user's profie directory, same as the Windows
USERPROFILE environment variable.