Fix broken out-of-tree build.
Reported by Marco Wahl. When building in a separate directory,
the y.tab.h file isn't found because the current directory is
not searched by the compiler. It searches only the original
source tree and the local config directory.

* Makefile (TXR_CFLAGS): We must have . in the include
file search path otherwise files can only find y.tab.h
when it's in the same directory where they are.