build: defend against $(top_srcdir) interference.
* Makefile (tainted): New top target added in a situation
when an out-of-tree build is tainted by a configuration
in the original tree. The problem is that .c files which
have an #include "y.tab.h" or #include "config.h" get
these headers from the same directory where they reside.
This is processed ahead of any -iquote directories.
If a "config.h" exists in $(top_srcdir), files include
that in spite of "-iquote ." on the compiler command line.