Bugfix: .1 treated as dot if preceded by space.
Some recent work in supporting .slot syntax (uref dot) broke
the treatment of floating point literals. This is because part
of the trick is that a uref dot is recognized with leading
whitespace as part of the token. But that of course means
it steals the match for some floating-point tokens; oops!

* parser.l (grammar): All rules for floating-point tokens
which can match a leading decimal point now munch optional
whitespace first.