doc: clarify behavior of mutated keys in hashing.
* txr.1: Under description of the equal method, replace the
text about a hash table "not working reliably" with proper
"unspecified behavior" terminology. Add some paragraphs to the
introduction to hash tables about the issue of modifying the
car fields of hash entry cells, or mutating the keys in
equal-based hash tables such that their equality is affected.
Re-iterate some of these points in a few other places under
the descriptions of some hash-related functions.