apply and iapply bugfix: split sequences into args.
These functions don't conform with the documentation.  For
instance [apply list "abc"] yields "abc".  It is supposed to
yield (#\a #\b #\c), since the characters of "abc" must become
individual arguments to list.  Part of the fix is in the
apply_frob_args logic; however, we have to clone that function
because it is used for implementing other things which
will break: we cannot, for for example, have (list* 1 "ab")
producing (1 #\a #\b).

* eval.c (apply_intrisic_frob_args): New static function.
Differs from apply_frob_args in that it calls tolist
on the final element.
(apply_intrinsic): Use apply_intrinsic_frob_args
instead of apply_frob_args.
(iapply): Invoke tolist on the value assigned to last_arg.

* txr.1: Add a clarifying note for iapply that the terminating
atom is not split into arguments if it is a sequence.