Indexing in variable subst applies to any sequence.
The @{a [3]} syntax in quasiliterals and @(output) now
indexes into the original object a if it is any sequence
kind, not specifically a list. Otherwise it indexes into
its string representation.

* eval.c (format_field): Combine the elements of the
object with the separator if it is any sequence type other
than a string.  Subject to compat option.
(subst_vars): Avoid converting any kind of sequence to
string, rather than just lists.  After any field formatting
is applied, if the object is any sequence (not just alist),
combine the elements with a space.  All subect to compat
option.

* match.c (tx_subst_vars): Same treatment as subst_vars.

* txr.1: Compatibility notes added.