Bugfix: missing usr:end and usr:single symbols.
When TXR pattern language code is canned in the context of
an alternative *package* which uses the usr package,
the scanner will throw false errors on @(end) and @(single).
This is because these symbols don't exist in the usr
package and will get interned in the current package.
The scanner will then complain that they are not in the
usr package.

* match.c (syms_init): Intern end and single in the
usr package, but don't store the return value in any
variable.