Makefile: detect when y.tab.h doesn't change.
When we touch the parser.y file, yacc regenerates
y.tab.h. Then everything which depends on that is
rebuilt. However, this is not necessary if the parser.y
change doesn't actually modify the contents of y.tab.h.
A build system based on file hashes wouldn't have this
problem, but make works strictly with modification
timestamps.

* Makefile (y.tab.c): Preserve the old y.tab.h, if
it exists, as y.tab.h.old. Then if the new y.tab.h
is exactly the same according to cmp, restore the
old y.tab.h in its place.
(clean): Remove y.tab.h.old.