ffi: buf type doesn't need fill function.
The fill function is useless, because the object is already filled
directly. It's just performing an exactly overlapping memcpy.

* ffi.c (ffi_buf_alloc): Function removed.
(ffi_type_compile): We can't use the presence or absence of
the fill function as the test whether a type can be passed
by ptr-in-out or ptr-out, since buf can be passed that way
and has no fill. A better criterion is "has no fill and
use the fixed size allocation".  Removing the assignment
statements which set up the fill function for buffers.