ffi: support anonymous padding in structs.
If the ffi struct type operator specifies nil as a slot name,
then we treat the member as anonymous padding. When reading a
struct from a buffer, we skip the padding. When writing out a
struct to a buffer, we generate zeros in the size of the
anonymous member. The type of the member is ignored, except
for calculating the size.

* ffi.c (ffi_struct_put): If the slot name is nil, don't
recursively call put; just use memset to fill the target
with zero bytes.
(ffi_struct_get, ffi_struct_fill): If the slot name is
nil, don't call get to fetch data. Just skip the area.