ffi: do arg in semantics /after/ getting return val.
What this allows is for situations when a foreign function
returns the pointer that it has been passed. If that pointer
is temporary storage allocated by FFI, then it is no longer
valid after performing the in pass on the args. Therefore, we
should decode the return value first, while the returned
pointer is valid.

* ffi.c (ffi_call_wrap): Move the return value get before
the argument post-processing in pass.