protsym: further refinement.
In sysif.c, we have conditionally defined symbol variables.
This has to be reflected with #if directivec in protsym.c.

* genprotsym.txr: Associate each symbol with a list
of preprocessor symbols which controls it, via a hash.
All symbols coming from a conditional file are tagged with
a list which includes at least the controlling symbol.
If symbols are wrapped with #if <sym> directives, that
is added. Nesting is supported. We group the symbols
according to their controlling preprocessor symbol lists.