ffi: add two tests based on glob function.
One approach captures the paths as a carray of
strings, and explicitly frees it with globfree.
The other approach uses a zarray, taking advantage
of null termination. globfree is elided because TXR FFI
does the freeing; the types used declare to it that it
is taking ownership of a dynamically allocated vector of
dynamically allocated strings, and so it performs the
equivalent of globfree.

* tests/017/glob-carray.expected: New file.

* tests/017/glob-carray.tl: Likewise.

* tests/017/glob-zarray.expected: Likewise.

* tests/017/glob-zarray.tl: Likewise.